% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_balance}
\alias{wkex_balance}
\title{The free balance of an asset.}
\usage{
wkex_balance(ex, asset)
}
\arguments{
\item{ex}{A `wickra_exchange` object.}

\item{asset}{Asset symbol, e.g. "BTC".}
}
\value{
The free balance as a number.
}
\description{
The free balance of an asset.
}
