% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_cancel_batch}
\alias{wkex_cancel_batch}
\title{Cancel several orders on a market in one request.}
\usage{
wkex_cancel_batch(adv, market, order_ids)
}
\arguments{
\item{adv}{A `wickra_advanced` object.}

\item{market}{Market symbol.}

\item{order_ids}{Character vector of venue order ids.}
}
\description{
Cancel several orders on a market in one request.
}
