% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_name}
\alias{wkex_name}
\title{The venue identifier of an exchange.}
\usage{
wkex_name(ex)
}
\arguments{
\item{ex}{A `wickra_exchange` object.}
}
\value{
The venue name ("paper", "replay", "binance", ...).
}
\description{
The venue identifier of an exchange.
}
