% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_place_limit}
\alias{wkex_place_limit}
\title{Place a limit order.}
\usage{
wkex_place_limit(ex, market, side, quantity, price)
}
\arguments{
\item{ex}{A `wickra_exchange` object.}

\item{market}{Market string.}

\item{side}{"buy" or "sell".}

\item{quantity}{Order quantity.}

\item{price}{Limit price.}
}
\value{
The resulting order as a list.
}
\description{
Place a limit order.
}
