% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_place_market}
\alias{wkex_place_market}
\title{Place a market order.}
\usage{
wkex_place_market(ex, market, side, quantity)
}
\arguments{
\item{ex}{A `wickra_exchange` object.}

\item{market}{Market string.}

\item{side}{"buy" or "sell".}

\item{quantity}{Order quantity.}
}
\value{
The resulting order as a list.
}
\description{
Place a market order.
}
