% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_poll}
\alias{wkex_poll}
\title{Drain buffered events.}
\usage{
wkex_poll(ex, capacity = 16L)
}
\arguments{
\item{ex}{A `wickra_exchange` object.}

\item{capacity}{Maximum events to return per call.}
}
\value{
A list of event lists.
}
\description{
Drain buffered events.
}
