% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_set_leverage}
\alias{wkex_set_leverage}
\title{Set the leverage for a market.}
\usage{
wkex_set_leverage(deriv, market, leverage)
}
\arguments{
\item{deriv}{A `wickra_derivatives` object.}

\item{market}{Market symbol.}

\item{leverage}{Integer leverage.}
}
\description{
Set the leverage for a market.
}
