% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_set_margin_mode}
\alias{wkex_set_margin_mode}
\title{Set the margin mode ("cross" or "isolated") for a market.}
\usage{
wkex_set_margin_mode(deriv, market, mode)
}
\arguments{
\item{deriv}{A `wickra_derivatives` object.}

\item{market}{Market symbol.}

\item{mode}{"cross" or "isolated".}
}
\description{
Set the margin mode ("cross" or "isolated") for a market.
}
