% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_subscribe_ticker}
\alias{wkex_subscribe_ticker}
\title{Subscribe to the ticker stream for a market.}
\usage{
wkex_subscribe_ticker(ex, market)
}
\arguments{
\item{ex}{A `wickra_exchange` object.}

\item{market}{Market symbol, e.g. "BTC/USDT".}
}
\value{
Invisibly, `ex`.
}
\description{
Subscribe to the ticker stream for a market.
}
