% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_subscribe_user_data}
\alias{wkex_subscribe_user_data}
\title{Open the private user-data stream.}
\usage{
wkex_subscribe_user_data(ud)
}
\arguments{
\item{ud}{A `wickra_user_data` object.}
}
\value{
Invisibly, `ud`.
}
\description{
Open the private user-data stream.
}
