% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_ticker}
\alias{wkex_ticker}
\title{The current ticker for a market.}
\usage{
wkex_ticker(ex, market)
}
\arguments{
\item{ex}{A `wickra_exchange` object.}

\item{market}{Market symbol, e.g. "BTC/USDT".}
}
\value{
A list with `symbol`, `last`, `bid`, `ask`, `volume`.
}
\description{
The current ticker for a market.
}
