% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{wkex_user_data_poll}
\alias{wkex_user_data_poll}
\title{Drain buffered user-data events.}
\usage{
wkex_user_data_poll(ud, capacity = 16L)
}
\arguments{
\item{ud}{A `wickra_user_data` object.}

\item{capacity}{Maximum events to return per call.}
}
\value{
A list of event lists.
}
\description{
Drain buffered user-data events.
}
